modelos que faltaban y algun seeder

+++ database/migrations/2022_01_27_113453_create_rating_respuestas_table.php
@@ -17,7 +17,7 @@ class CreateRatingRespuestasTable extends Migration
         Schema::create('rating_respuestas', function (Blueprint $table) {
             $table->foreignId('user_id')->constrained();
             $table->foreignId('respuesta_id')->constrained();
-            $table->integer('rating');
+            $table->smallInteger('rating');
             $table->timestamps();
             $table->primary(['user_id', 'respuesta_id']);
         });
